Tests: Extract TickUtils.spec.ts
